初步对接接口

+++ src/main/java/buaa/jj/accountservice/api/IUserService.java
@@ -3,10 +3,10 @@ package buaa.jj.accountservice.api;
 import java.util.Date;
 public interface IUserService {
     public int Recharge(int UserID,String RequestTime,String RequestID,
-                        float Money,boolean Way,boolean RequestState);
+                        double Money,boolean Way,boolean RequestState);
     public int Withdraw(int UserID,String RequestTime,String RequestID,
-                        float Money,boolean Way,boolean RequestState);
+                        double Money,boolean Way,boolean RequestState);
     public int Consume(int UserID,int MerchantId,String RequestTime,String RequestID,
-                       float Money,boolean RequestState);
+                       double Money,boolean RequestState);
 }
